requirements.txt
requests == 2.32.3
pyyaml >= 6.0.1
argparse >= 1.4.0
beautifulsoup4 >= 4.12.3
lxml >= 5.2.2
